Changed the REST class to do the filtering in a seperate function. This is later being used in other code parts that need to, for any reason, sanitize the URL. By default, if no arguments are provided, it uses the data send by PHP.
Added a "setUrl" function to the core that uses the url sanitize filter from the REST class.
Added a "getUrl" function to the core that retreives the data created by setUrl. It returns an array with
the url als SOE url [seo] and with the url as argument based url [arg].