Major infrastructure change. An increasingly bigger issue became the inability to let a module read data from another module. This has now been added and in the progress completely obseleted the "AdminHelper" class since that where all special things for modules which are now nicely added to the module named: "modules". This also made the Core get/set functions somewhat vague since the module data can't be retreived by the get/set function but only by the getModuleData and setModuleData. For now this remains the case, but this might change in the future. Also added is a "soft" check to see if the new data class is existing within a module. It's not a hard requirement but it's being checked when adding a module or even when looking at the module overview.
Now what does this infrastructure change allow you to do? Simple, you can now make a module, let it manage it's special data part and use it in that module. However another module can now also use that data as long as you added it in the MODULE_data class in the data.php file, For example. If there is one module that adds a logging ability then all other modules can make use of the log module to log their stuff.